requirement.txt
# requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
supabase==2.0.3

# ML/AI
sentence-transformers==2.2.2
onnxruntime==1.16.3
transformers==4.35.2
torch==2.1.1
numpy==1.24.3

# Media Processing
yt-dlp==2023.11.16
opencv-python==4.8.1.78
easyocr==1.7.0
moviepy==1.0.3
imageio-ffmpeg==0.4.9

# Audio Processing
openai-whisper==20231117
librosa==0.10.1
soundfile==0.12.1

# HTTP Client
httpx==0.25.1

# Utils
pydantic==2.5.0
